requirements/base.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
sqlparse==0.4.2
python-dotenv~=0.19.2
python-slugify      # https://github.com/un33k/python-slugify
Pillow              # https://github.com/python-pillow/Pillow
redis               # https://github.com/redis/redis-py
hiredis             # https://github.com/redis/hiredis-py
celery              # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat  # https://github.com/celery/django-celery-beat
flower              # https://github.com/mher/flower
uvicorn[standard]   # https://github.com/encode/uvicorn
environ
django-environ
channels
channels_redis

# Django
# ------------------------------------------------------------------------------
django==4.0.1  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.47.0  # https://github.com/pennersr/django-allauth
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.11.0 # https://github.com/adamchainz/django-cors-headers